doc: bag interface, bag union, unit tests syntax

+++ genc/prelude.go
@@ -127,7 +127,7 @@ ku_print(str s) {
 		if (n < 0) {
 			return;
 		}
-		i += n;
+		i += (uint)(n);
 	}
 }
